% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pt_municip.R
\name{map_pt_municip}
\alias{map_pt_municip}
\title{Plot Municipalities in Portugal}
\usage{
map_pt_municip(
  data,
  main_var,
  municip_var,
  pal = NULL,
  coordinates = list(x_continent = c(-12, -6), y_continent = c(36.7, 42), x_madeira =
    c(-17.5, -16), y_madeira = c(32.5, 33.2), x_acores = c(-31, -25.5), y_acores = c(37,
    39.5)),
  map = ptmap::pt_municip,
  map_europe = ptmap::europe,
  extra = NULL
)
}
\arguments{
\item{data}{A tibble with the municipality level information and a minicipality code column.}

\item{main_var}{The variable with the municipality level info. E.g., median wage by municipality. Not a string.}

\item{municip_var}{Unique ID of the Municipality}

\item{pal}{A palette to be used in the colors of the plot}

\item{coordinates}{This list contains the coordinates for mainland portugal, madeira and a?ores.}

\item{map}{What is the base map that should be used.}

\item{map_europe}{An attempt to draw Spain}
}
\value{
A ggplot object with the map. The theme can still be changed and other options/geoms added.
}
\description{
This function creates a map with information for Portugal at the municipality level.
The function automatically places the islands next to mainland Portugal.
}
\examples{
data(ine_municip_wage)

ine_municip_wage \%>\% 
  map_pt_municip(median_wage,
                 cod)

}
\author{
Miguel Salema
}
