% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pt_nuts_II.R
\name{map_pt_nuts_II}
\alias{map_pt_nuts_II}
\title{Plot NUTS2 in Portugal}
\usage{
map_pt_nuts_II(
  data,
  main_var,
  nut2_var,
  pal = NULL,
  coordinates = list(x_continent = c(-12, -6), y_continent = c(36.7, 42), x_madeira =
    c(-17.5, -16), y_madeira = c(32.5, 33.2), x_acores = c(-31, -25.5), y_acores = c(37,
    39.5)),
  map = ptmap::europe,
  nut_var_qp = FALSE
)
}
\arguments{
\item{data}{A tibble with the nut2 level information and a nut2 column.}

\item{main_var}{The variable with the NUT2 level info. E.g., average wage by nut. Not a string.}

\item{nut2_var}{Name of the column that has the info about the nuts. Not a string.}

\item{pal}{A palette to be used in the colors of the plot}

\item{coordinates}{This list contains the coordinates for mainland portugal, madeira and a?ores.}

\item{map}{What is the base map that should be used.}

\item{nut_var_qp}{If true, that means that the NUT2 classification in the variable is not correct, but yes as in my QP cleaning (1, 2,...). If this is TRUE, it automatically transforms this variable.}
}
\value{
A ggplot object with the map. The theme can still be changed and other options/geoms added.
}
\description{
This function creates a map with information for Portugal at the NUT2 level.
The function automatically places the islands next to mainland Portugal.
It has the feature nut_var_qp, that should be TRUE if the user is using QP data with the cleaning of Miguel Salema.
}
\examples{
library(dplyr)
data(qp_nut2_data) # load wage data from Portugal at NUT2 level

# create a tibble with only one year, 2019.
df19 <- qp_nut2_data \%>\%
  filter(year == 2019)

# Apply the function
map_pt_19 <- df19 \%>\%
  map_pt_nuts_II(main_var = mean_real_wage,
                 nut2_var = nut_2_est,
                 nut_var_qp = T)

# The output is a ggplot map that can be easily altered
map_pt_19 +
  labs(fill = "Real Average Wage") # changes the title of the legend

# we can put 2 graphs next to each other
df93 <- qp_nut2_data \%>\%
  filter(year == 1993)

map_pt_93 <- df93 \%>\%
  map_pt_nuts_II(main_var = mean_real_wage,
                 nut2_var = nut_2_est,
                 nut_var_qp = TRUE)

library(gridExtra)
grid.arrange(map_pt_93,
             map_pt_19,
             ncol=2)

}
\author{
Miguel Salema
}
